[refactor] address, category, jwt 도메인 리팩토링 (#109)
* refactor: Address Dto 디렉토리를 Member 하위에서 Address 도메인 하위로 이동

* refactor: CategoryReadResponse 생성자 접근 제어자를 private으로 변경

* style: 주석 추가

* feat: CategoryException의 기본 에러 메시지 추가

* refactor: Entity에서 @AllArgsConstructor 생성자 제거

* refactor: CategoryReadResponse의 정적 팩토리 내부 로직을 빌더 패턴으로 변경
