+++ be/src/main/java/kr/codesquad/secondhand/api/category/dto/CategoryReadResponse.java
@@ -3,6 +3,7 @@ package kr.codesquad.secondhand.api.category.dto;
 import java.util.List;
 import java.util.stream.Collectors;
 import kr.codesquad.secondhand.api.category.domain.Category;
+import lombok.Builder;
 import lombok.Getter;
 
 @Getter
@@ -12,19 +13,28 @@ public class CategoryReadResponse {
     private final String name;
     private final String imgUrl;
 
-    public CategoryReadResponse(Long id, String name, String imgUrl) {
+    @Builder
+    private CategoryReadResponse(Long id, String name, String imgUrl) {
         this.id = id;
         this.name = name;
         this.imgUrl = imgUrl;
     }
 
     public static CategoryReadResponse from(Category category) {
-        return new CategoryReadResponse(category.getId(), category.getName(), category.getImgUrl());
+        return CategoryReadResponse.builder()
+                .id(category.getId())
+                .name(category.getName())
+                .imgUrl(category.getImgUrl())
+                .build();
     }
 
     public static List<CategoryReadResponse> from(List<Category> categories) {
         return categories.stream()
-                .map(category -> new CategoryReadResponse(category.getId(), category.getName(), category.getImgUrl()))
+                .map(category -> CategoryReadResponse.builder()
+                        .id(category.getId())
+                        .name(category.getName())
+                        .imgUrl(category.getImgUrl())
+                        .build())
                 .collect(Collectors.toUnmodifiableList());
     }
 }
