+++ be/src/main/java/kr/codesquad/secondhand/api/jwt/domain/Jwt.java
@@ -1,13 +1,15 @@
 package kr.codesquad.secondhand.api.jwt.domain;
 
-import lombok.AllArgsConstructor;
 import lombok.Getter;
 
 @Getter
-@AllArgsConstructor
 public class Jwt {
 
-    private String accessToken;
-    private String refreshToken;
+    private final String accessToken;
+    private final String refreshToken;
 
+    public Jwt(String accessToken, String refreshToken) {
+        this.accessToken = accessToken;
+        this.refreshToken = refreshToken;
+    }
 }
