+++ be/src/main/java/kr/codesquad/secondhand/api/jwt/domain/MemberRefreshToken.java
@@ -5,14 +5,12 @@ import javax.persistence.Entity;
 import javax.persistence.Id;
 import javax.persistence.Table;
 import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Entity
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
 @Table(name = "token")
 public class MemberRefreshToken {
 
@@ -20,6 +18,11 @@ public class MemberRefreshToken {
     private Long memberId;
 
     private String refreshToken;
+
+    public MemberRefreshToken(Long memberId, String refreshToken) {
+        this.memberId = memberId;
+        this.refreshToken = refreshToken;
+    }
 
     public boolean matches(String refreshToken) {
         return Objects.equals(this.refreshToken, refreshToken);
